requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.29.0
pydantic[email]~=2.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv
pydantic-settings
Jinja2
SQLAlchemy~=2.0
pandas~=2.2.0
scikit-learn~=1.5.0
numpy==1.26.4
psycopg2-binary~=2.9